fix up scripts

+++ Test.ps1
@@ -2,6 +2,7 @@ param (
     [string]$Platform = "x64",
     [string]$Configuration = "debug",
     [switch]$IsAzurePipelineBuild = $false,
+    [switch]$RunUITests = $false,
     [switch]$Help = $false
 )
 
@@ -89,27 +90,23 @@ if (-not (Test-Path -Path "AppxPackages")) {
 try {
     foreach ($platform in $env:Build_Platform.Split(",")) {
         foreach ($configuration in $env:Build_Configuration.Split(",")) {
-            $DevHomePackage = Get-AppPackage "Microsoft.DevHome" -ErrorAction SilentlyContinue
-            if ($DevHomePackage) {
-                Write-Host "Uninstalling old Dev Home"
-                Remove-AppPackage -Package $DevHomePackage.PackageFullName
-            }
-            Write-Host "Installing Dev Home"
-            Add-AppPackage (Join-Path "AppxPackages" "$configuration\DevHome-$platform.msix")
-
-            if ($true) {
-                # Start/stop the app once so that WinAppDriver doesn't time out
-                Start-Process "Shell:AppsFolder\Microsoft.Windows.DevHome.Dev_8wekyb3d8bbwe!App"
-                for ($i - 0; $i -lt 60; $i++) {
-                    $temp = Get-Process -Name "DevHome" -ea SilentlyContinue
-                    if ($temp -eq $Null) {
-                        Write-Host "null"
-                    } else {
-                        Write-host $temp.Id
-                    }
-                    Start-Sleep 1
+            # TODO: UI tests are currently disabled in the pipeline until signing is solved
+            if ($RunUITests) {
+                $DevHomePackage = Get-AppPackage "Microsoft.DevHome" -ErrorAction SilentlyContinue
+                if ($DevHomePackage) {
+                    Write-Host "Uninstalling old Dev Home"
+                    Remove-AppPackage -Package $DevHomePackage.PackageFullName
                 }
-                Stop-Process -Name "DevHome"
+                Write-Host "Installing Dev Home"
+                Add-AppPackage (Join-Path "AppxPackages" "$configuration\DevHome-$platform.msix")
+
+                if ($true) {
+                    # Start/stop the app once so that WinAppDriver doesn't time out during first time setup
+                    # and wait 60 seconds to give plenty of time
+                    Start-Process "Shell:AppsFolder\Microsoft.Windows.DevHome.Dev_8wekyb3d8bbwe!App"
+                    Start-Sleep 60
+                    Stop-Process -Name "DevHome"
+                }
             }
 
             $vstestArgs = @(
@@ -124,7 +121,33 @@ try {
                 "uitest\bin\$platform\$configuration\net8.0-windows10.0.22621.0\DevHome.UITest.dll"
             )
 
-            & $vstestPath $winAppTestArgs
+            & $vstestPath $vstestArgs
+            # TODO: UI tests are currently disabled in the pipeline until signing is solved
+            if ($RunUITests) {
+                & $vstestPath $winAppTestArgs
+            }
+
+            foreach ($toolPath in (Get-ChildItem "tools")) {
+                $tool = $toolPath.Name
+                $vstestArgs = @(
+                    "/Platform:$platform",
+                    "/Logger:trx;LogFileName=$tool.Test-$platform-$configuration.trx",
+                    "tools\$tool\*UnitTest\bin\$platform\$configuration\net8.0-windows10.0.22621.0\*.UnitTest.dll"
+                )
+
+                $winAppTestArgs = @(
+                    "/Platform:$platform",
+                    "/Logger:trx;LogFileName=$tool.UITest-$platform-$configuration.trx",
+                    "/Settings:uitest\Test.runsettings",
+                    "tools\$tool\*UITest\bin\$platform\$configuration\net8.0-windows10.0.22621.0\*.UITest.dll"
+                )
+
+                & $vstestPath $vstestArgs
+                # TODO: UI tests are currently disabled in the pipeline until signing is solved
+                if ($RunUITests) {
+                    & $vstestPath $winAppTestArgs
+                }
+            }
         }
     }
 } catch {
